ProjectPanels made.
Receipts need to be worked out still also added an action listener to
the shop panel and made the source for the store a constant string
instead of being passed into the constructor. Updated main frame to
create the project panels. updated projectManager to add functionality
needed for the project panels. Made list classes implement serializable
in order to have them saved with the projects also changed them back to
arrayList class because that implements serializable and list doesn't.